Giving size_t type argument an initial value causes problem.
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@9215 3895a583-e809-0410-9bb6-aa9463026377

+++ FEM/DistributionInfo.cpp
@@ -40,7 +40,7 @@ FiniteElement::DistributionType DistributionInfo::getProcessDistributionType () 
 
   WW 24.08.2011
 */
-LinearFunctionData::LinearFunctionData(ifstream &ins, const size_t num_var)
+LinearFunctionData::LinearFunctionData(ifstream &ins, const int num_var)
 {
    string str_buff;
    vector<string> tokens;
